Remove unnecessary anonymous namespaces
